+++ CHANGELOG.md
@@ -3,6 +3,29 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.5.0] - 2020-04-21
+
+### Changes
+
+- Bio:
+  - Add: Bio image
+  - Add: Social links
+  - Fix: many iOS glitches fixed
+
+## [0.4.0] - 2020-04-20
+
+### Changes
+
+- Layout:
+  - Add: Background
+
+## [0.3.1] - 2020-04-20
+
+### Changes
+
+- Layout:
+  - Add: Remove animations on mobile
+
 ## [0.3.0] - 2020-04-20
 
 ### Changes